refactor(structure): reorganiza carpetas core, setup y ui para mayor claridad
- Renombrado `core/` a `engine/`, siguiendo convenciones en gamedev.
- Movido `core/game/` a `engine/game/`.
- Movido `entities/setup/` a nueva carpeta `loaders/`.
- Reorganizada carpeta `ui/` en `ui/menus/` y `ui/widgets/`.
- Agregados archivos `__init__.py` en todos los directorios con código para definir paquetes Python.
