+++ roguelike_project/engine/game/systems/entity_loader.py
@@ -1,7 +1,7 @@
 from roguelike_project.entities.player.base import Player
-from roguelike_project.entities.setup.load_obstacles import load_obstacles
-from roguelike_project.entities.setup.load_enemies import load_enemies
-from roguelike_project.entities.setup.load_buildings import load_buildings
+from roguelike_project.loaders.load_obstacles import load_obstacles
+from roguelike_project.loaders.load_enemies import load_enemies
+from roguelike_project.loaders.load_buildings import load_buildings
 
 def load_entities():
     player = Player(600, 600)
