+++ roguelike_project/loaders/load_obstacles.py
